Unit_Converter/requirements.txt
flask
flask-cors
flasgger
requests
kafka-python   
Flask-SQLAlchemy  
mysql-connector-python                                                                                          